+++ CHANGELOG.md
@@ -88,6 +88,7 @@
 - [10527](https://github.com/vegaprotocol/vega/issues/10527) - Add support for `byte32` type in market proposal oracle definition.
 - [10563](https://github.com/vegaprotocol/vega/issues/10563) - Spam protection for create/update referral program.
 - [10246](https://github.com/vegaprotocol/vega/issues/10246) - Add quantum volumes to teams statistics API.
+- [10550](https://github.com/vegaprotocol/vega/issues/10550) - Update network parameters with default values.
 
 ### 🐛 Fixes